Arreglo test de controllers reclamo y talonario
